server/requirements.txt
asgiref==3.11.1
dj-database-url==3.1.2
Django==6.0.2
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-yasg==1.21.15
inflection==0.5.1
packaging==26.0
psycopg==3.3.3
psycopg-binary==3.3.3
PyJWT==2.11.0
python-dotenv==1.2.2
pytz==2025.2
PyYAML==6.0.3
redis==6.4.0
sqlparse==0.5.5
uritemplate==4.2.0
